Move script generation to another module.
