+++ tests/test_environment_scripts.py
@@ -5,7 +5,7 @@ import sys
 
 import pytest
 
-from conda_devenv.devenv import render_activate_script, render_deactivate_script
+from conda_devenv.gen_scripts import render_activate_script, render_deactivate_script
 
 
 @pytest.fixture
